Revert "防止一些基本类被继承，提高代码安全性"
This reverts commit aa41c3911403b2bbd1a8cd5aff26189e6664ded3.

+++ command/command.cpp
@@ -246,7 +246,7 @@ namespace qqbot
 
 					Command::splitCommand(taskName, commandName, commandArgs);
 
-					timers.addTask(taskName, static_cast<int>(interval), &Command::groupExcute, this, groupID, senderID, commandName, commandArgs);
+					timers.addTask(taskName, interval, &Command::groupExcute, this, groupID, senderID, commandName, commandArgs);
 
 					Network::sendGroupMessage(groupID, std::format("添加\"{}\"指令到定时器中成功，间隔为{}毫秒", taskName, interval));
 				}
